requirements.txt
boto3==1.12.39
botocore==1.15.39
docutils==0.15.2
gitdb==4.0.4
GitPython==3.1.1
jmespath==0.9.5
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.14.0
smmap==3.0.2
urllib3==1.25.8
flask==1.1.2
